Support KMP in sample + detect in plugin
