--- sample/src/main/kotlin/dev/zacsweers/redacted/sample/User.kt
@@ -1,9 +1,0 @@
-package dev.zacsweers.redacted.sample
-
-import dev.zacsweers.redacted.annotations.Redacted
-
-data class User(
-    val name: String,
-    @Redacted val phoneNumber: String
-)
-
